Add tests for different scenarios.
Homesick will ask to overwrite a file/symlink if it impedes the creation
of the directory structure. If you chose not to overwrite, it will
follow the symlink, and continue symlinking castle files in this new
directory.

Homesick will detect if a file in a castle is identical, in cases where
a directory in the homedir is symlinked to a castle directory. (It
compares if the absolute paths are identical).