Conform to the XDG Base Directory Specification
The XDG Base Directory Specification[0] aims to provide guidelines for
organizing configuration, data and cache files for a system. Some of the
benefits include a less cluttered home directory, easy backups and
predictability. This commit changes the default directory for homsick
castles to be `XDG_DATA_HOME/homesick` instead of `~/.homesick`.

[0]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html